requirements.txt
bleak
paho-mqtt
pydantic